+++ Library/gltfLoader/tiny_gltf.h
@@ -7222,19 +7222,19 @@ static void WriteBinaryGltfStream(std::ostream &stream,
       return numToRound + multiple - remainder;
   };
 
-  const uint32_t padding_size = roundUp(content.size(), 4) - content.size();
+  const uint32_t padding_size = roundUp((uint32_t)content.size(), 4) - (uint32_t)content.size();
 
   // 12 bytes for header, JSON content length, 8 bytes for JSON chunk info.
   // Chunk data must be located at 4-byte boundary.
-  const int length = 12 + 8 + roundUp(content.size(), 4)+
-      (binBuffer.size()?(8+roundUp(binBuffer.size(),4)) : 0);
+  const int length = 12 + 8 + roundUp((UINT)content.size(), 4)+
+      ((UINT)binBuffer.size()?(8+roundUp((UINT)binBuffer.size(),4)) : 0);
 
   stream.write(header.c_str(), std::streamsize(header.size()));
   stream.write(reinterpret_cast<const char *>(&version), sizeof(version));
   stream.write(reinterpret_cast<const char *>(&length), sizeof(length));
 
   // JSON chunk info, then JSON data
-  const int model_length = int(content.size()) + padding_size;
+  const int model_length = int((uint32_t)content.size()) + padding_size;
   const int model_format = 0x4E4F534A;
   stream.write(reinterpret_cast<const char *>(&model_length),
                sizeof(model_length));
@@ -7248,9 +7248,9 @@ static void WriteBinaryGltfStream(std::ostream &stream,
     stream.write(padding.c_str(), std::streamsize(padding.size()));
   }
   if (binBuffer.size() > 0){
-    const uint32_t bin_padding_size = roundUp(binBuffer.size(), 4) - binBuffer.size();
+    const uint32_t bin_padding_size = roundUp((UINT)binBuffer.size(), 4) - (UINT)binBuffer.size();
     // BIN chunk info, then BIN data
-    const int bin_length = int(binBuffer.size()) + bin_padding_size;
+    const int bin_length = int((UINT)binBuffer.size()) + bin_padding_size;
     const int bin_format = 0x004e4942;
     stream.write(reinterpret_cast<const char *>(&bin_length),
 		 sizeof(bin_length));
